fix stuffz

+++ Makefile
@@ -4,7 +4,8 @@
 .PHONY: \
 	clean \
 	clean-disk-images \
-	clean-tmp
+	clean-tmp \
+	vm
 
 clean: postgres-kill postgres-clean dev-clean geodata-clean clean-tmp clean-disk-images
 
@@ -12,8 +13,7 @@ clean-disk-images:
 	rm -f *.qcow2
 
 clean-tmp: postgres-clean
-	rm -rf $$(readlink ./tmp)
-	unlink ./tmp
+	if [ -L ./tmp ]; then rm -rf $$(readlink ./tmp); unlink ./tmp; fi
 
 ./tmp:
 	ln -sf $$(mktemp -d) ./tmp
